requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make lock
#
aiohttp-security==0.4.0   # via -r requirements/base.txt
aiohttp-session==2.9.0    # via -r requirements/base.txt
aiohttp==3.6.2            # via -r requirements/base.txt, aiohttp-security, aiohttp-session, jibrel-aiohttp-swagger, pytest-aiohttp
aiopg==1.0.0              # via -r requirements/base.txt
async-timeout==3.0.1      # via -r requirements/base.txt, aiohttp
attrs==19.3.0             # via -r requirements/base.txt, aiohttp, pytest
cffi==1.14.0              # via -r requirements/base.txt, cryptography
chardet==3.0.4            # via -r requirements/base.txt, aiohttp
click==7.1.1              # via -r requirements/base.txt, pip-tools
cryptography==2.9         # via -r requirements/base.txt
entrypoints==0.3          # via flake8
factory-boy==2.12.0       # via -r requirements/dev.in, pytest-factoryboy
faker==0.9.1              # via factory-boy
flake8==3.7.9             # via -r requirements/dev.in
idna==2.9                 # via -r requirements/base.txt, yarl
inflection==0.4.0         # via pytest-factoryboy
isort==4.3.21             # via -r requirements/dev.in
jibrel-aiohttp-swagger==0.1.0  # via -r requirements/base.txt
loguru==0.4.1             # via -r requirements/base.txt
marshmallow==3.5.1        # via -r requirements/base.txt, webargs
mccabe==0.6.1             # via flake8
more-itertools==8.2.0     # via pytest
multidict==4.7.5          # via -r requirements/base.txt, aiohttp, yarl
mypy-extensions==0.4.3    # via mypy
mypy==0.770               # via -r requirements/dev.in
packaging==20.3           # via pytest
passlib==1.7.2            # via -r requirements/base.txt
pip-tools==4.5.1          # via -r requirements/dev.in
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.5    # via -r requirements/base.txt, aiopg
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8
pycparser==2.20           # via -r requirements/base.txt, cffi
pyflakes==2.1.1           # via flake8
pyparsing==2.4.7          # via packaging
pytest-aiohttp==0.3.0     # via -r requirements/dev.in
pytest-factoryboy==2.0.3  # via -r requirements/dev.in
pytest==5.4.1             # via -r requirements/dev.in, pytest-aiohttp, pytest-factoryboy
python-dateutil==2.8.1    # via faker
pyyaml==5.3.1             # via -r requirements/base.txt, jibrel-aiohttp-swagger
six==1.14.0               # via -r requirements/base.txt, cryptography, faker, packaging, pip-tools, python-dateutil
sqlalchemy==1.3.16        # via -r requirements/base.txt
text-unidecode==1.2       # via faker
typed-ast==1.4.1          # via mypy
typing-extensions==3.7.4.2  # via mypy
wcwidth==0.1.9            # via pytest
webargs==6.0.0            # via -r requirements/base.txt
yarl==1.4.2               # via -r requirements/base.txt, aiohttp
